feat: split up Signatures::CreateSignFlow
The logic for signer selection of a regular piece vs a report or the
minutes is different enough that it makes more sense to split these off
entirely.
