--- app/components/signatures/create-sign-flow.js
@@ -1,216 +1,0 @@
-import Component from '@glimmer/component';
-import { tracked } from '@glimmer/tracking';
-import { action } from '@ember/object';
-import { inject as service } from '@ember/service';
-import { task } from 'ember-concurrency';
-import { TrackedArray } from 'tracked-built-ins';
-import { trackedFunction } from 'ember-resources/util/function';
-import { startOfDay } from 'date-fns';
-import CONSTANTS from 'frontend-kaleidos/config/constants';
-
-/**
- */
-export default class SignaturesCreateSignFlowComponent extends Component {
-  @service mandatees;
-  @service store;
-
-  @tracked showSecretaryModal = false;
-  @tracked showMinisterModal = false;
-  @tracked showApproversModal = false;
-  @tracked showNotificationAddressesModal = false;
-
-  @tracked signers = new TrackedArray([]);
-  @tracked approvers = new TrackedArray([]);
-  @tracked notificationAddresses = new TrackedArray([]);
-
-  @tracked hasConflictingSigners = false;
-  primeMinister = null;
-
-  constructor() {
-    super(...arguments);
-  }
-
-  loadSigners = trackedFunction(this, async () => {
-    if (!this.args.decisionActivities) {
-      return;
-    }
-    const decisionActivities = this.args.decisionActivities.toArray();
-
-    const [head, ...tail] = decisionActivities;
-    let submitter;
-    let cosigners = [];
-
-    if (this.args.isDecisionOrNota) {
-      for (let decisionActivity of tail) {
-        if (head.secretary?.id !== decisionActivity.secretary?.id) {
-          this.hasConflictingSigners = true;
-          break;
-        }
-      }
-    } else {
-      this.primeMinister = await this.store.queryOne('mandatee', {
-        'filter[mandate][role][:uri:]':
-          CONSTANTS.MANDATE_ROLES.MINISTER_PRESIDENT,
-        'filter[:lt:start]': startOfDay(new Date()).toISOString(),
-        'filter[:has-no:end]': true,
-        include: 'person,mandate.role',
-      });
-
-      const getSubmitterAndCosigners = async (decisionActivity) => {
-        const subcase = await decisionActivity.subcase;
-        const submitter = await subcase.requestedBy;
-        const cosigners = await subcase.mandatees;
-        return { submitter, cosigners };
-      };
-
-      const { submitter, cosigners } = await getSubmitterAndCosigners(head);
-
-      for (let decisionActivity of tail) {
-        const { submitter: _submitter, cosigners: _cosigners } =
-          await getSubmitterAndCosigners(decisionActivity);
-
-        if (submitter?.id !== _submitter?.id) {
-          this.hasConflictingSigners = true;
-          break;
-        }
-        if (_cosigners.length !== cosigners.length) {
-          this.hasConflictingSigners = true;
-          break;
-        }
-        const _cosignersIds = _cosigners.map((signer) => signer.id);
-        const cosignersIds = cosigners.map((signer) => signer.id);
-        for (const mandateeId of _cosignersIds) {
-          if (!cosignersIds.includes(mandateeId)) {
-            this.hasConflictingSigners = true;
-            break;
-          }
-        }
-        if (this.hasConflictingSigners) break;
-      }
-    }
-    if (this.hasConflictingSigners) {
-      this.signers = new TrackedArray([]);
-    } else {
-      const signersSet = new Set();
-      if (this.args.isDecisionOrNota) {
-        const secretary = await head.secretary;
-        signersSet.add(secretary);
-      } else {
-        signersSet.add(this.primeMinister);
-        const primeMinisterPerson = await this.primeMinister.person;
-
-        if (submitter) {
-          const submitterPerson = await submitter.person;
-          if (primeMinisterPerson.id !== submitterPerson.id) {
-            signersSet.add(submitter);
-          }
-        }
-
-        for (let cosigner of cosigners) {
-          if (cosigner) {
-            const cosignerPerson = await cosigner.person;
-            if (primeMinisterPerson.id !== cosignerPerson.id) {
-              signersSet.add(cosigner);
-            }
-          }
-        }
-
-      }
-      this.signers = new TrackedArray([...signersSet]);
-    }
-    this.args.onChangeSigners?.(this.signers);
-  });
-
-  @action
-  startEditSigners() {
-    if (this.args.isDecisionOrNota) {
-      this.showSecretaryModal = true;
-    } else {
-      this.showMinisterModal = true;
-    }
-  }
-
-  @action
-  saveApprover(approver) {
-    this.approvers.addObject(approver);
-    this.showApproversModal = false;
-    this.args.onChangeApprovers?.(this.approvers);
-  }
-
-  @action
-  removeApprover(approver) {
-    this.approvers.removeObject(approver);
-    this.args.onChangeApprovers?.(this.approvers);
-  }
-
-  @action
-  saveNotificationAddress(address) {
-    this.notificationAddresses.addObject(address);
-    this.showNotificationAddressesModal = false;
-    this.args.onChangeNotificationAddresses?.(this.notificationAddresses);
-  }
-
-  @action
-  removeNotificationAddress(address) {
-    this.notificationAddresses.removeObject(address);
-    this.args.onChangeNotificationAddresses?.(this.notificationAddresses);
-  }
-
-  saveSigners = task(async (selected) => {
-    console.log(selected);
-    if (this.args.isDecisionOrNota) {
-      const decisionActivities = this.args.decisionActivities.toArray();
-      for (let decisionActivity of decisionActivities) {
-        const secretary = await decisionActivity.secretary;
-        if (secretary.id !== selected.id) {
-          decisionActivity.secretary = selected;
-          await decisionActivity.save();
-        }
-      }
-      this.signers = new Set([selected])
-      this.showSecretaryModal = false;
-    } else {
-      const records = await this.mandateeIdsToRecords(selected);
-      const filtered = await this.filterSelectedSigners(records);
-      this.signers = filtered.sort((m1, m2) => m1.priority - m2.priority);
-      this.showMinisterModal = false;
-    }
-    
-    this.args.onChangeSigners?.(this.signers);
-  });
-
-  @action
-  removeSigner(signer) {
-    this.signers.removeObject(signer);
-    this.args.onChangeSigners?.(this.signers);
-  }
-
-  async mandateeIdsToRecords(mandatees) {
-    return Promise.all(
-      mandatees.map((id) => this.store.findRecord('mandatee', id))
-    );
-  }
-
-  /**
-   * Filter the selected ministers for signing so that it only contains
-   * the selected mandatees except for the Minister-President's non-PM
-   * mandatee.
-   */
-  async filterSelectedSigners(selected) {
-    const filtered = new Set();
-    for (const mandatee of selected) {
-      if (this.primeMinister) {
-        const primeMinisterPerson = await this.primeMinister.person;
-        const person = await mandatee.person;
-        if (person.id === primeMinisterPerson.id) {
-          if (mandatee.id === this.primeMinister.id) {
-            filtered.add(this.primeMinister);
-          }
-          continue;
-        }
-      }
-      filtered.add(mandatee);
-    }
-    return [...filtered];
-  }
-}
