+++ app/components/signatures/create-sign-flow/piece.hbs
@@ -1,0 +1,203 @@
+{{#if this.hasConflictingSigners}}
+  <AuAlert
+    @skin="warning"
+    @icon="alert-triangle"
+    @title={{t "create-sign-flow-conflicting-signers-title"}}
+  >
+    <p>{{t "create-sign-flow-conflicting-signers-message"}}</p>
+  </AuAlert>
+{{/if}}
+<Auk::Panel class="auk-u-mt-2">
+  <Auk::Panel::Header class="auk-u-px-1">
+    <Auk::Toolbar as |Toolbar|>
+      <Toolbar.Group @position="left" as |Group|>
+        <Group.Item>
+          <h4 class="auk-h4 auk-u-m-0">
+            {{t "to-sign-by"}}
+          </h4>
+        </Group.Item>
+      </Toolbar.Group>
+      <Toolbar.Group @position="right" as |Group|>
+        <Group.Item>
+          <AuButton
+            data-test-signatures-create-sign-flow-signers-edit
+            @skin="secondary"
+            @icon="pencil"
+            @loading={{this.loadSigners.isPending}}
+            {{on "click" this.startEditSigners}}
+          >
+            {{t "edit-list"}}
+          </AuButton>
+        </Group.Item>
+      </Toolbar.Group>
+    </Auk::Toolbar>
+  </Auk::Panel::Header>
+  <Auk::Panel::Body class="auk-u-p-0">
+    <AuList
+      @divider={{true}}
+      as |Item|
+    >
+      {{#if this.loadSigners.isPending}}
+        <Item class="au-u-padding-small">
+          <Auk::Loader />
+        </Item>
+      {{else}}
+        {{#each this.signers as |signer|}}
+          <Item
+            class="auk-o-flex auk-o-flex--justify-between auk-o-flex--vertical-center au-u-padding-small"
+            data-test-signatures-create-sign-flow-signers-item
+          >
+            {{signer.person.fullName}}
+            <AuButton
+              data-test-signatures-create-sign-flow-signers-remove
+              @skin="naked"
+              @alert={{true}}
+              @hideText={{true}}
+              @icon="trash"
+              {{on "click" (fn this.removeSigner signer)}}
+            >
+              {{t "delete"}}
+            </AuButton>
+          </Item>
+        {{else}}
+          <Item class="au-u-padding-small">
+            {{t "no-signers-selected"}}
+          </Item>
+        {{/each}}
+      {{/if}}
+    </AuList>
+  </Auk::Panel::Body>
+</Auk::Panel>
+<Auk::Panel class="auk-u-mt-4">
+  <Auk::Panel::Header class="auk-u-px-1">
+    <Auk::Toolbar as |Toolbar|>
+      <Toolbar.Group @position="left" as |Group|>
+        <Group.Item>
+          <h4 class="auk-h4 auk-u-m-0">
+            {{t "to-approve-by"}}
+          </h4>
+        </Group.Item>
+      </Toolbar.Group>
+      <Toolbar.Group @position="right" as |Group|>
+        <Group.Item>
+          <AuButton
+            data-test-signatures-create-sign-flow-approvers-add
+            @skin="secondary"
+            @icon="plus"
+            {{on "click" (toggle "showApproversModal" this)}}
+          >
+            {{t "add"}}
+          </AuButton>
+        </Group.Item>
+      </Toolbar.Group>
+    </Auk::Toolbar>
+  </Auk::Panel::Header>
+  <Auk::Panel::Body class="auk-u-p-0">
+    <AuList
+      @divider={{true}}
+      as |Item|
+    >
+      {{#each this.approvers as |approver|}}
+        <Item
+          class="auk-o-flex auk-o-flex--justify-between auk-o-flex--vertical-center au-u-padding-small"
+          data-test-signatures-create-sign-flow-approvers-item
+        >
+          <Auk::AbbreviatedText @text={{approver}} @maxLength={{35}} />
+          <AuButton
+            data-test-signatures-create-sign-flow-approvers-remove
+            @skin="naked"
+            @alert={{true}}
+            @hideText={{true}}
+            @icon="trash"
+            {{on "click" (fn this.removeApprover approver)}}
+          >
+            {{t "delete"}}
+          </AuButton>
+        </Item>
+      {{else}}
+        <Item class="au-u-padding-small">
+          {{t "no-approvers-selected"}}
+        </Item>
+      {{/each}}
+    </AuList>
+  </Auk::Panel::Body>
+</Auk::Panel>
+<Auk::Panel class="auk-u-mt-4">
+  <Auk::Panel::Header class="auk-u-px-1">
+    <Auk::Toolbar as |Toolbar|>
+      <Toolbar.Group @position="left" as |Group|>
+        <Group.Item>
+          <h4 class="auk-h4 auk-u-m-0">
+            {{t "signing-hub-cc"}}
+          </h4>
+        </Group.Item>
+      </Toolbar.Group>
+      <Toolbar.Group @position="right" as |Group|>
+        <Group.Item>
+          <AuButton
+            data-test-signatures-create-sign-flow-notification-adresses-add
+            @skin="secondary"
+            @icon="plus"
+            {{on "click" (toggle "showNotificationAddressesModal" this)}}
+          >
+            {{t "add"}}
+          </AuButton>
+        </Group.Item>
+      </Toolbar.Group>
+    </Auk::Toolbar>
+  </Auk::Panel::Header>
+  <Auk::Panel::Body class="auk-u-p-0">
+    <AuList
+      @divider={{true}}
+      as |Item|
+    >
+      {{#each this.notificationAddresses as |notificationAddress|}}
+        <Item
+          class="auk-o-flex auk-o-flex--justify-between auk-o-flex--vertical-center au-u-padding-small"
+          data-test-signatures-create-sign-flow-notification-adresses-item
+        >
+          <Auk::AbbreviatedText @text={{notificationAddress}} @maxLength={{35}} />
+          <AuButton
+            data-test-signatures-create-sign-flow-notification-adresses-remove
+            @skin="naked"
+            @alert={{true}}
+            @hideText={{true}}
+            @icon="trash"
+            {{on "click" (fn this.removeNotificationAddress notificationAddress)}}
+          >
+            {{t "delete"}}
+          </AuButton>
+        </Item>
+      {{else}}
+        <Item class="au-u-padding-small">
+          {{t "no-notification-addresses-selected"}}
+        </Item>
+      {{/each}}
+    </AuList>
+  </Auk::Panel::Body>
+</Auk::Panel>
+
+{{#if this.showMinisterModal}}
+  <Signatures::SelectMinistersModal
+    @loading={{this.saveSigners.isRunning}}
+    @selected={{map-by "id" this.signers}}
+    @onConfirm={{this.saveSigners.perform}}
+    @onCancel={{toggle "showMinisterModal" this}}
+  />
+{{/if}}
+
+{{#if this.showApproversModal}}
+  <Signatures::EmailModal
+    @title={{t "add-approval"}}
+    @onConfirm={{this.saveApprover}}
+    @onCancel={{toggle "showApproversModal" this}}
+  />
+{{/if}}
+
+{{#if this.showNotificationAddressesModal}}
+  <Signatures::EmailModal
+    @title={{t "add-notification"}}
+    @onConfirm={{this.saveNotificationAddress}}
+    @onCancel={{toggle "showNotificationAddressesModal" this}}
+  />
+{{/if}}
